@@ -11,25 +11,25 @@
     "package:firefox": "web-ext build -s dist-firefox --overwrite-dest"
   },
   "dependencies": {
-    "esbuild": "^0.14.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "webextension-polyfill": "^0.8.0"
+    "esbuild": "^0.15.10",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "webextension-polyfill": "^0.10.0"
   },
   "devDependencies": {
-    "@types/node": "^17.0.2",
-    "@types/react-dom": "^17.0.11",
-    "@types/webextension-polyfill": "^0.8.2",
-    "@typescript-eslint/eslint-plugin": "^5.8.0",
-    "@typescript-eslint/parser": "^5.8.0",
-    "chokidar": "^3.5.2",
-    "eslint": "^8.5.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "prettier": "2.5.1",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.4",
-    "web-ext": "^6.6.0"
+    "@types/node": "^18.8.2",
+    "@types/react-dom": "^18.0.6",
+    "@types/webextension-polyfill": "^0.9.1",
+    "@typescript-eslint/eslint-plugin": "^5.39.0",
+    "@typescript-eslint/parser": "^5.39.0",
+    "chokidar": "^3.5.3",
+    "eslint": "^8.24.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-react": "^7.31.8",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "2.7.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.4",
+    "web-ext": "^7.2.0"
   }
 }
